requirements.txt
anyio==3.6.2
asgiref==3.5.2
astroid==2.12.13
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
autopep8==2.0.0
case-converter==1.1.0
certifi==2022.9.24
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
databases==0.6.2
Deprecated==1.2.13
dill==0.3.6
distlib==0.3.6
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
exceptiongroup==1.0.4
fastapi==0.87.0
fastapi-pagination==0.11.1
filelock==3.8.0
flake8==5.0.4
h11==0.14.0
httptools==0.5.0
identify==2.5.9
idna==3.4
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
limits==1.6
MarkupSafe==2.1.1
mccabe==0.7.0
mock==4.0.3
msal==1.20.0
nodeenv==1.7.0
orjson==3.8.2
packaging==21.3
platformdirs==2.5.4
pluggy==1.0.0
pre-commit==2.20.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
py==1.11.0
pyaml-env==1.2.0
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pyflakes==2.5.0
PyJWT==2.6.0
pylint==2.15.6
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
redis==4.3.4
requests==2.28.1
rsa==4.9
six==1.16.0
slowapi==0.1.6
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.21.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.16.7
watchfiles==0.18.1
watchgod==0.8.2
websockets==10.4
wrapt==1.14.1
